backend/requirements.txt
Flask==2.2.3 
gunicorn==20.1.0
Werkzeug==2.2.2
requests==2.28.2
python-dotenv==0.10.1
pydantic==2.6.3
Flask-Pydantic==0.12.0
scikit-learn==1.4.1.post1
networkx==3.2.1
pandas==2.2.0
numpy==1.26.4